Change colors to int rather than uint8_t
Change colors to int to allow for special manipulations
like value wrapping. Also add center parameter for
set_target for drawing centered textures.

+++ src/color.h
@@ -7,8 +7,8 @@
 
 namespace Dynamo {
     struct Color {
-        uint8_t r, g, b; 
-        uint8_t a = 255; // Default alpha value
+        int r, g, b; 
+        int a = 255; // Default alpha value
 
         // Convert color to hexidecimal
         uint32_t get_hex();
